fix : Entity and Repository & finish adminComment form

+++ app/public/index.php
@@ -102,10 +102,15 @@ try {
         $request = $_GET['admin-comment'];
         if ($request == 'list') {
             $adminCommentController->index();
-        } elseif ($request == 'delete' && key_exists('id_comment', $_GET)) {
-            $adminCommentController->delete($_GET['id_comment']);
         } elseif ($request == 'toggle') {
             $adminCommentController->toggle();
+        } elseif (
+            $request == 'delete'
+            && key_exists('_method', $_POST)
+            && key_exists('id_comment', $_GET)
+            && ($_POST['_method'] == "DELETE")
+        ) {
+            $adminCommentController->delete($_GET['id_comment']);
         }
     } else {
         $contactController = new ContactController($twig, $userRepository);
